fixes nested signup and login components
